Replace manual ioctl calls with nix crate
ifr in new_from_fd was not marked mut. This was a bug since ioctl call
modified this variable. This lead to optimizer in 1.85.0 version of Rust
to optimize away the access to this variable after ioctl call and just
storing hardcoded 0 in the flags variable.
This commit fixes it and additionally:
1. Uses ioctl macro from nix crate
2. Removes unnecesary struct definitions which are available in libc
   crate